autoradar/requirements.txt
# Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
pydantic==2.10.4
pydantic-settings==2.7.1

# Scraping
playwright>=1.49.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Telegram
python-telegram-bot==21.9

# Scheduling
apscheduler==3.10.4

# Utilities
httpx==0.28.1
python-dotenv==1.0.1
